+++ R/poi.R
@@ -9,8 +9,9 @@
 #' polygon, mimicking the behavior of [graphics::polypath()].
 #'
 #' @param x a vector of x coordinates or a matrix or data.frame
-#'   of x and y coordinates, a list of components x and y, or
-#'   a time series (see [grDevices::xy.coords()] for details)
+#'   of x and y coordinates, a list of components x and y,
+#'   a time series (see [grDevices::xy.coords()] for details),
+#'   or a simple features object from package `sf`.
 #' @param y a vector of y coordinates. Only needs to be provided if
 #'   `x` is vector.
 #' @param precision the precision to use when computing the center
@@ -46,7 +47,21 @@
 #'
 #' points(poi(x, y))
 #'
+#' \dontrun{
+#' # Find visual centers for North Carolina counties
+#' library(sf)
+#' nc <- st_read(system.file("shape/nc.shp", package="sf"))
+#' locations = do.call(rbind, poi(nc, precision=0.01))
+#' plot(st_geometry(nc))
+#' points(locations)
+
+#' }
 poi <- function(x, y = NULL, precision = 1.0) {
+  UseMethod('poi')
+}
+
+#' @export
+poi.default <- function(x, y = NULL, precision = 1.0) {
   xy <- grDevices::xy.coords(x, y)
 
   stopifnot(is.numeric(precision),