+++ NEWS.md
@@ -1,5 +1,9 @@
 # polylabelr 0.1.0.9000
 
+## New features
+
+* `poi()` supports simple features objects from the `sf` package (#2, thanks @kent37).
+
 ## Bug fixes
 
 * The `CITATION` file now correctly displays the year of the package.